Passing S800 map on command line.
A file with ".inv" extension will be assumed to be the S800 map.
This will be loaded the first time an TS800 object is instantiated.
If no ".inv" file is passed, it will try to load "invmap.inv".

A mutex is used to ensure that this does not cause issues if two threads
  instantiate a TS800 at the same time.

Closes #37.